Remove v1 node api code
