+++ api/Cargo.toml
@@ -15,7 +15,6 @@ failure = "0.1.1"
 failure_derive = "0.1.1"
 hyper = "0.13"
 lazy_static = "1"
-regex = "1"
 ring = "0.16"
 serde = "1"
 serde_derive = "1"
@@ -23,12 +22,10 @@ serde_json = "1"
 log = "0.4"
 tokio = { version = "0.2", features = ["full"] }
 tokio-rustls = "0.13"
-http = "0.2"
 hyper-rustls = "0.20"
 hyper-timeout = "0.3"
 futures = "0.3"
 rustls = "0.17"
-url = "2.1"
 bytes = "0.5"
 
 grin_core = { path = "../core", version = "5.0.0-beta.2" }
