+++ api/src/error.rs
@@ -1,0 +1,91 @@
+// Copyright 2020 The Grin Developers
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Errors that can be returned by an ApiEndpoint implementation.
+use crate::router::RouterError;
+use failure::{Backtrace, Context, Fail};
+use std::fmt::{self, Display};
+
+#[derive(Debug)]
+pub struct Error {
+	inner: Context<ErrorKind>,
+}
+
+#[derive(Clone, Eq, PartialEq, Debug, Fail, Serialize, Deserialize)]
+pub enum ErrorKind {
+	#[fail(display = "Internal error: {}", _0)]
+	Internal(String),
+	#[fail(display = "Bad arguments: {}", _0)]
+	Argument(String),
+	#[fail(display = "Not found.")]
+	NotFound,
+	#[fail(display = "Request error: {}", _0)]
+	RequestError(String),
+	#[fail(display = "ResponseError error: {}", _0)]
+	ResponseError(String),
+	#[fail(display = "Router error: {}", _0)]
+	Router(RouterError),
+}
+
+impl Fail for Error {
+	fn cause(&self) -> Option<&dyn Fail> {
+		self.inner.cause()
+	}
+
+	fn backtrace(&self) -> Option<&Backtrace> {
+		self.inner.backtrace()
+	}
+}
+
+impl Display for Error {
+	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+		Display::fmt(&self.inner, f)
+	}
+}
+
+impl Error {
+	pub fn kind(&self) -> &ErrorKind {
+		self.inner.get_context()
+	}
+}
+
+impl From<ErrorKind> for Error {
+	fn from(kind: ErrorKind) -> Error {
+		Error {
+			inner: Context::new(kind),
+		}
+	}
+}
+
+impl From<Context<ErrorKind>> for Error {
+	fn from(inner: Context<ErrorKind>) -> Error {
+		Error { inner: inner }
+	}
+}
+
+impl From<RouterError> for Error {
+	fn from(error: RouterError) -> Error {
+		Error {
+			inner: Context::new(ErrorKind::Router(error)),
+		}
+	}
+}
+
+impl From<crate::chain::Error> for Error {
+	fn from(error: crate::chain::Error) -> Error {
+		Error {
+			inner: Context::new(ErrorKind::Internal(error.to_string())),
+		}
+	}
+}
