+++ api/src/handlers/server_api.rs
@@ -14,32 +14,14 @@
 
 use super::utils::w;
 use crate::chain::{Chain, SyncState, SyncStatus};
+use crate::error::*;
 use crate::p2p;
-use crate::rest::*;
-use crate::router::{Handler, ResponseFuture};
 use crate::types::*;
-use crate::web::*;
-use hyper::{Body, Request};
 use serde_json::json;
 use std::convert::TryInto;
 use std::sync::Weak;
 
-// RESTful index of available api endpoints
-// GET /v1/
-pub struct IndexHandler {
-	pub list: Vec<String>,
-}
-
-impl IndexHandler {}
-
-impl Handler for IndexHandler {
-	fn get(&self, _req: Request<Body>) -> ResponseFuture {
-		json_response_pretty(&self.list)
-	}
-}
-
 /// Status handler. Post a summary of the server status
-/// GET /v1/status
 pub struct StatusHandler {
 	pub chain: Weak<Chain>,
 	pub peers: Weak<p2p::Peers>,
@@ -64,12 +46,6 @@ impl StatusHandler {
 			api_sync_status,
 			api_sync_info,
 		))
-	}
-}
-
-impl Handler for StatusHandler {
-	fn get(&self, _req: Request<Body>) -> ResponseFuture {
-		result_to_response(self.get_status())
 	}
 }
