+++ api/src/router.rs
@@ -293,86 +293,3 @@ fn collect_node_middleware(handlers: &mut Vec<HandlerObj>, node: &Node) {
 		}
 	}
 }
-
-#[cfg(test)]
-mod tests {
-
-	use super::*;
-	use futures::executor::block_on;
-
-	struct HandlerImpl(u16);
-
-	impl Handler for HandlerImpl {
-		fn get(&self, _req: Request<Body>) -> ResponseFuture {
-			let code = self.0;
-			Box::pin(async move {
-				let res = Response::builder()
-					.status(code)
-					.body(Body::default())
-					.unwrap();
-				Ok(res)
-			})
-		}
-	}
-
-	#[test]
-	fn test_add_route() {
-		let mut routes = Router::new();
-		let h1 = Arc::new(HandlerImpl(1));
-		let h2 = Arc::new(HandlerImpl(2));
-		let h3 = Arc::new(HandlerImpl(3));
-		routes.add_route("/v1/users", h1.clone()).unwrap();
-		assert!(routes.add_route("/v1/users", h2.clone()).is_err());
-		routes.add_route("/v1/users/xxx", h3.clone()).unwrap();
-		routes.add_route("/v1/users/xxx/yyy", h3.clone()).unwrap();
-		routes.add_route("/v1/zzz/*", h3.clone()).unwrap();
-		assert!(routes.add_route("/v1/zzz/ccc", h2.clone()).is_err());
-		routes
-			.add_route("/v1/zzz/*/zzz", Arc::new(HandlerImpl(6)))
-			.unwrap();
-	}
-
-	#[test]
-	fn test_get() {
-		let mut routes = Router::new();
-		routes
-			.add_route("/v1/users", Arc::new(HandlerImpl(101)))
-			.unwrap();
-		routes
-			.add_route("/v1/users/xxx", Arc::new(HandlerImpl(103)))
-			.unwrap();
-		routes
-			.add_route("/v1/users/xxx/yyy", Arc::new(HandlerImpl(103)))
-			.unwrap();
-		routes
-			.add_route("/v1/zzz/*", Arc::new(HandlerImpl(103)))
-			.unwrap();
-		routes
-			.add_route("/v1/zzz/*/zzz", Arc::new(HandlerImpl(106)))
-			.unwrap();
-
-		let call_handler = |url| {
-			let task = async {
-				let resp = routes
-					.get(url)
-					.unwrap()
-					.next()
-					.unwrap()
-					.get(Request::new(Body::default()))
-					.await
-					.unwrap();
-				resp.status().as_u16()
-			};
-			block_on(task)
-		};
-
-		assert_eq!(call_handler("/v1/users"), 101);
-		assert_eq!(call_handler("/v1/users/xxx"), 103);
-		assert!(routes.get("/v1/users/yyy").is_err());
-		assert_eq!(call_handler("/v1/users/xxx/yyy"), 103);
-		assert!(routes.get("/v1/zzz").is_err());
-		assert_eq!(call_handler("/v1/zzz/1"), 103);
-		assert_eq!(call_handler("/v1/zzz/2"), 103);
-		assert_eq!(call_handler("/v1/zzz/2/zzz"), 106);
-	}
-}
