+++ api/src/web.rs
@@ -1,4 +1,4 @@
-use crate::rest::*;
+use crate::error::*;
 use crate::router::ResponseFuture;
 use bytes::Buf;
 use futures::future::ok;
@@ -6,9 +6,7 @@ use hyper::body;
 use hyper::{Body, Request, Response, StatusCode};
 use serde::{Deserialize, Serialize};
 use serde_json;
-use std::collections::HashMap;
 use std::fmt::Debug;
-use url::form_urlencoded;
 
 /// Parse request body
 pub async fn parse_body<T>(req: Request<Body>) -> Result<T, Error>
@@ -21,27 +19,6 @@ where
 
 	serde_json::from_reader(raw.bytes())
 		.map_err(|e| ErrorKind::RequestError(format!("Invalid request body: {}", e)).into())
-}
-
-/// Convert Result to ResponseFuture
-pub fn result_to_response<T>(res: Result<T, Error>) -> ResponseFuture
-where
-	T: Serialize,
-{
-	match res {
-		Ok(s) => json_response_pretty(&s),
-		Err(e) => match e.kind() {
-			ErrorKind::Argument(msg) => response(StatusCode::BAD_REQUEST, msg.clone()),
-			ErrorKind::RequestError(msg) => response(StatusCode::BAD_REQUEST, msg.clone()),
-			ErrorKind::NotFound => response(StatusCode::NOT_FOUND, ""),
-			ErrorKind::Internal(msg) => response(StatusCode::INTERNAL_SERVER_ERROR, msg.clone()),
-			ErrorKind::ResponseError(msg) => {
-				response(StatusCode::INTERNAL_SERVER_ERROR, msg.clone())
-			}
-			// place holder
-			ErrorKind::Router(_) => response(StatusCode::INTERNAL_SERVER_ERROR, ""),
-		},
-	}
 }
 
 /// Utility to serialize a struct into JSON and produce a sensible Response
@@ -81,106 +58,3 @@ pub fn just_response<T: Into<Body> + Debug>(status: StatusCode, text: T) -> Resp
 pub fn response<T: Into<Body> + Debug>(status: StatusCode, text: T) -> ResponseFuture {
 	Box::pin(ok(just_response(status, text)))
 }
-
-pub struct QueryParams {
-	params: HashMap<String, Vec<String>>,
-}
-
-impl QueryParams {
-	pub fn process_multival_param<F>(&self, name: &str, mut f: F)
-	where
-		F: FnMut(&str),
-	{
-		if let Some(ids) = self.params.get(name) {
-			for id in ids {
-				for id in id.split(',') {
-					f(id);
-				}
-			}
-		}
-	}
-
-	pub fn get(&self, name: &str) -> Option<&String> {
-		self.params.get(name).and_then(|v| v.first())
-	}
-}
-
-impl From<&str> for QueryParams {
-	fn from(query_string: &str) -> Self {
-		let params = form_urlencoded::parse(query_string.as_bytes())
-			.into_owned()
-			.fold(HashMap::new(), |mut hm, (k, v)| {
-				hm.entry(k).or_insert_with(|| vec![]).push(v);
-				hm
-			});
-		QueryParams { params }
-	}
-}
-
-impl From<Option<&str>> for QueryParams {
-	fn from(query_string: Option<&str>) -> Self {
-		match query_string {
-			Some(query_string) => Self::from(query_string),
-			None => QueryParams {
-				params: HashMap::new(),
-			},
-		}
-	}
-}
-
-impl From<Request<Body>> for QueryParams {
-	fn from(req: Request<Body>) -> Self {
-		Self::from(req.uri().query())
-	}
-}
-
-#[macro_export]
-macro_rules! right_path_element(
-	($req: expr) =>(
-		match $req.uri().path().trim_end_matches('/').rsplit('/').next() {
-			None => return response(StatusCode::BAD_REQUEST, "invalid url"),
-			Some(el) => el,
-		};
-	));
-
-#[macro_export]
-macro_rules! must_get_query(
-	($req: expr) =>(
-		match $req.uri().query() {
-			Some(q) => q,
-			None => return Err(ErrorKind::RequestError("no query string".to_owned()).into()),
-		}
-	));
-
-#[macro_export]
-macro_rules! parse_param(
-	($param: expr, $name: expr, $default: expr) =>(
-	match $param.get($name) {
-		None => $default,
-		Some(val) =>  match val.parse() {
-			Ok(val) => val,
-			Err(_) => return Err(ErrorKind::RequestError(format!("invalid value of parameter {}", $name)).into()),
-		}
-	}
-	));
-
-#[macro_export]
-macro_rules! parse_param_no_err(
-	($param: expr, $name: expr, $default: expr) =>(
-	match $param.get($name) {
-		None => $default,
-		Some(val) =>  match val.parse() {
-			Ok(val) => val,
-			Err(_) => $default,
-		}
-	}
-	));
-
-#[macro_export]
-macro_rules! w_fut(
-	($p: expr) =>(
-		match w($p) {
-			Ok(p) => p,
-			Err(_) => return response(StatusCode::INTERNAL_SERVER_ERROR, "weak reference upgrade failed" ),
-		}
-	));
